Add color utilities tests
