+++ app.html
@@ -115,6 +115,7 @@
     <button id="showUiFab" class="fab" hidden data-i18n="btnShowPanel" title="显示面板">显示面板</button>
     <div id="toast" role="status" aria-live="polite"></div>
 
+    <script src="color-utils.js"></script>
     <script src="app.js"></script>
   </body>
   </html>
